Move "Dependency" and "NodeDesc" to "typing".
