+++ pylib-v1/ailets/cons/typing.py
@@ -1,5 +1,44 @@
-from typing import Optional, Protocol
+from dataclasses import dataclass
+from typing import Optional, Protocol, Sequence
 from .streams import Stream
+
+
+@dataclass(frozen=True)
+class Dependency:
+    """A dependency of a node on another node's stream.
+
+    Attributes:
+        name: Optional name to reference this dependency in the node's inputs
+        source: Name of the node this dependency comes from
+        stream: Optional name of the specific stream from the source node
+    """
+
+    source: str
+    name: Optional[str] = None
+    stream: Optional[str] = None
+
+    def to_json(self) -> list:
+        """Convert to JSON-serializable format.
+
+        Returns:
+            List of [dep_name, node_name, stream_name]
+        """
+        return [self.name, self.source, self.stream]
+
+    @classmethod
+    def from_json(cls, data: list) -> "Dependency":
+        """Create dependency from JSON data.
+
+        Args:
+            data: List of [dep_name, node_name, stream_name]
+        """
+        return cls(name=data[0], source=data[1], stream=data[2])
+
+
+@dataclass(frozen=True)
+class NodeDesc:
+    name: str
+    inputs: Sequence[Dependency]
 
 
 class IEnvironment(Protocol):
